backend/requirements.txt
fastapi
ultralytics
opencv-python
uvicorn
python-multipart
supabase
python-jose[cryptography]
python-dotenv
